Changed api view in another file
